+++ src/demo/clear-global-value-params/clear-global-value-params.js
@@ -2,7 +2,7 @@ import Adtrace from '../../sdk/main'
 import SimpleAction from '../simple-action'
 
 function init () {
-  SimpleAction('cleargpp', Adtrace.clearGlobalPartnerParameters)()
+  SimpleAction('cleargpp', Adtrace.clearGlobalValueParameters)()
 }
 
 export default init