+++ src/demo/remove-global-value-param/remove-global-value-param.js
@@ -1,7 +1,7 @@
 import RemoveGlobalParam from '../remove-global-param'
 
 function init () {
-  RemoveGlobalParam('removegpp', 'removeGlobalPartnerParameter')()
+  RemoveGlobalParam('removegpp', 'removeGlobalValueParameter')()
 }
 
 export default init